Move test_derecho_mpirun to its own system testing file.
This is the last step needed to make it so that UNIT tests are portable to machines other than those defined in CESM, and thus:
Resolves ESCOMP/CTSM#2984.
