+++ python/ctsm/unit_testing.py
@@ -1,6 +1,21 @@
 """Functions to aid unit tests"""
 
+import sys
 from ctsm.ctsm_logging import setup_logging_for_tests
+
+
+def add_args(machine, nodes, tasks):
+    """add arguments to sys.argv"""
+    args_to_add = [
+        "--machine",
+        machine,
+        "--number-of-nodes",
+        str(nodes),
+        "--tasks-per-node",
+        str(tasks),
+    ]
+    for item in args_to_add:
+        sys.argv.append(item)
 
 
 def setup_for_tests(enable_critical_logs=False):
